Split GREU.jl into Model.jl, Calibrate.jl (merging with Calibration.jl), and Shock.jl
Calibrate saves the baseline to a parquet file. Shock reads the baseline from a parquet file.
