+++ model_julia/Calibrate.jl
@@ -1,9 +1,9 @@
-module Calibration
-
+# Calibrate the model: solve the baseline (static then dynamic), verify it, and export it
+# to Output/baseline.parquet for later use by Shock.jl.
 using SquareModels
-import ..Log: @log_time
-import ..Time: at_year, variable_year, t1, T
-import ..Tags: ForecastConstant
+include("Model.jl")
+import .Time: at_year, variable_year, t1, T
+import .Tags: ForecastConstant
 
 function residual_tolerances(data::ModelDictionary, submodels)
 	tolerances = ModelDictionary(data.model)
@@ -89,4 +89,51 @@ function calibrate_model(db, submodels)
 	return @log_time solve(block, db; replace_nothing=1.0)
 end
 
+# ==============================================================================
+# Solve calibration (static then dynamic)
+# ==============================================================================
+# Static: single-period at t1 — calibrates residuals and parameters
+Time.T = Settings.calibration_year
+@log_time static_solution = calibrate_model(db, submodels)
+@log_errors assert_residuals_small(
+	static_solution;
+	atol=1e-1,
+  tolerances=residual_tolerances(static_solution, submodels),
+	msg="Large residuals after static calibration"
+)
+
+# Dynamic: full horizon — uses static solution as starting values
+Time.T = Time.max_terminal_year
+@log_time baseline = calibrate_model(static_solution, submodels)
+@log_errors assert_residuals_small(
+	baseline;
+	atol=1e-1,
+  tolerances=residual_tolerances(baseline, submodels),
+	msg="Large residuals after dynamic calibration"
+)
+
+# ==============================================================================
+# Tests
+# ==============================================================================
+# Zero shock test: After calibration, solving the base model with no changes should give identical results
+@log_time begin
+	zero_shock = solve(base_model(), baseline)
+	assert_no_diff(baseline, zero_shock; atol=1e-6, msg="Zero shock test failed")
 end
+
+# Module-specific tests: collect failures from every module before raising, since a single
+# underlying bug often trips several checks at once (across one or more modules).
+test_errors = String[]
+for m in submodels
+	isdefined(m, :run_tests) && append!(test_errors, m.run_tests(baseline))
+end
+isempty(test_errors) || error(
+	"$(length(test_errors)) module test(s) failed:\n" * join(("  " * e for e in test_errors), "\n")
+)
+
+# ==============================================================================
+# Export baseline
+# ==============================================================================
+const output_dir = joinpath(@__DIR__, "..", "Output")
+mkpath(output_dir)
+@log_time unload(joinpath(output_dir, "baseline.parquet"), baseline)
